requirements.txt
requests
html2text
openai
python-dotenv